requirements.txt
PyQt6>=6.4.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pyecharts>=2.0.0 